server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "dotenv>=0.9.9",
    "langchain-chroma>=0.2.4",
    "langchain-ollama>=0.3.3",
    "langchain-text-splitters>=0.3.8",
    "pip>=25.1.1",
    "pymupdf>=1.26.0",
    "python-docx>=1.1.2",
    "spacy>=3.8.7",
    "tiktoken>=0.9.0",
]
